+++ controllers/clientRestController.js
@@ -39,7 +39,7 @@ let validatorMap = {
     "positiveDecimalAllowed": async function (keysToValidate, body) {
         let valid = {isValid: true, message: ""};
         for (var keyString of keysToValidate){
-            if (!body[keyString] || !Number.parseInt(body[keyString])
+            if (!body[keyString] || !Number.parseFloat(body[keyString])
                 || body[keyString].includes("-")){
                 valid.isValid = false;
                 valid.message += `${keyString} was not valid.  `;
